+++ phoenix-knowledge-api/src/main/java/com/ginkgocap/ywxt/knowledge/service/KnowledgeOtherService.java
@@ -13,13 +13,13 @@ import java.util.List;
  */
 public interface KnowledgeOtherService
 {
-    public InterfaceResult collectKnowledge(long userId,long knowledgeId, short columnId) throws Exception;
+    public InterfaceResult collectKnowledge(long userId,long knowledgeId, int columnId) throws Exception;
 
-    public InterfaceResult deleteCollectedKnowledge(long userId,long knowledgeId, short columnId) throws Exception;
+    public InterfaceResult deleteCollectedKnowledge(long userId,long knowledgeId, int columnId) throws Exception;
 
-    public boolean isCollectedKnowledge(long userId,long knowledgeId, short columnId);
+    public boolean isCollectedKnowledge(long userId,long knowledgeId, int columnId);
 
-    public List<KnowledgeCollect> myCollectKnowledge(long userId,short columnId,int start, int size) throws Exception;
+    public List<KnowledgeCollect> myCollectKnowledge(long userId,int columnId,int start, int size) throws Exception;
 
     public long myCollectKnowledgeCount(long userId) throws Exception;
